# Conflicts:
#	src/services/objects/StaticItemService.java

+++ src/resources/commands/Command.java
@@ -58,11 +58,11 @@ public class Command {
 	private int validWeapon;
 	private int invalidWeapon;
 	private String cooldownGroup;
-	private int warmupTime;
-	private int executeTime;
-	private int cooldownTime;
+	private float warmupTime;
+	private float executeTime;
+	private float cooldownTime;
 	private String cooldownGroup2;
-	private int cooldownTime2;
+	private float cooldownTime2;
 	// toolbarOnly
 	// fromServerOnly
 	private boolean autoAddToToolbar;
@@ -108,16 +108,16 @@ public class Command {
 	public void setInvalidWeapon(int invalidWeapon) { this.invalidWeapon = invalidWeapon; }
 	public String getCooldownGroup() { return cooldownGroup; }
 	public void setCooldownGroup(String cooldownGroup) { this.cooldownGroup = cooldownGroup; }
-	public int getWarmupTime() { return warmupTime; }
-	public void setWarmupTime(int warmupTime) { this.warmupTime = warmupTime; }
-	public int getExecuteTime() { return executeTime; }
-	public void setExecuteTime(int executeTime) { this.executeTime = executeTime; }
-	public int getCooldownTime() { return cooldownTime; }
-	public void setCooldownTime(int cooldownTime) { this.cooldownTime = cooldownTime; }
+	public float getWarmupTime() { return warmupTime; }
+	public void setWarmupTime(float warmupTime) { this.warmupTime = warmupTime; }
+	public float getExecuteTime() { return executeTime; }
+	public void setExecuteTime(float executeTime) { this.executeTime = executeTime; }
+	public float getCooldownTime() { return cooldownTime; }
+	public void setCooldownTime(float cooldownTime) { this.cooldownTime = cooldownTime; }
 	public String getCooldownGroup2() { return cooldownGroup2; }
 	public void setCooldownGroup2(String cooldownGroup2) { this.cooldownGroup2 = cooldownGroup2; }
-	public int getCooldownTime2() { return cooldownTime2; }
-	public void setCooldownTime2(int cooldownTime2) { this.cooldownTime2 = cooldownTime2; }
+	public float getCooldownTime2() { return cooldownTime2; }
+	public void setCooldownTime2(float cooldownTime2) { this.cooldownTime2 = cooldownTime2; }
 	public boolean isAutoAddToToolbar() { return autoAddToToolbar; }
 	public void setAutoAddToToolbar(boolean autoAddToToolbar) { this.autoAddToToolbar = autoAddToToolbar; }
